public/a2a_ge_oauth_example/backend/requirements.txt
a2a-sdk==0.2.12
google-adk==1.7.0
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.104.0
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-secret-manager==2.24.0
google-cloud-trace==1.16.2
google-genai==1.25.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
uvicorn==0.35.0
asyncpg==0.30.0
sqlalchemy==2.0.39
google-cloud-alloydb==0.4.8
google-cloud-alloydb-connector==1.9.0
greenlet==3.1.1
python-dotenv==1.0.0
nest-asyncio
pytz
